// vinta-school-os/src/types/settings.ts
export interface Academy {
  id: string
  name: string
  phone: string
  email: string
  address: string
  weekend_day: number
  current_term: string
  created_at: string
  updated_at: string
}